+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [31.0.2] - 2025-10-07
+
+Added option to hide share links
+
+
+## [31.0.1] - 2025-10-07
+
+Updated to latest Nextcloud
 
 ## [0.0.2] - 2017-07-31